+++ exchangesingle_test.go
@@ -5,11 +5,13 @@ import (
 	"encoding/json"
 	"errors"
 
-	"github.com/dogmatiq/dodeca/logging"
 	. "github.com/dogmatiq/harpy"
 	. "github.com/dogmatiq/harpy/internal/fixtures"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
+	"go.uber.org/zap/zaptest/observer"
 )
 
 var _ = Describe("func Exchange() (single request)", func() {
@@ -18,8 +20,8 @@ var _ = Describe("func Exchange() (single request)", func() {
 		request   Request
 		reader    *RequestSetReaderStub
 		writer    *ResponseWriterStub
-		buffer    *logging.BufferedLogger
-		logger    DefaultExchangeLogger
+		logs      *observer.ObservedLogs
+		logger    *ZapExchangeLogger
 		closed    bool
 	)
 
@@ -70,10 +72,11 @@ var _ = Describe("func Exchange() (single request)", func() {
 			},
 		}
 
-		buffer = &logging.BufferedLogger{}
+		var core zapcore.Core
+		core, logs = observer.New(zapcore.DebugLevel)
 
-		logger = DefaultExchangeLogger{
-			Target: buffer,
+		logger = &ZapExchangeLogger{
+			Target: zap.New(core),
 		}
 
 		closed = false
@@ -114,9 +117,16 @@ var _ = Describe("func Exchange() (single request)", func() {
 			)
 
 			Expect(err).ShouldNot(HaveOccurred())
-			Expect(buffer.Messages()).To(ContainElement(
-				logging.BufferedLogMessage{
-					Message: `call "<method>" [params: 2 B, result: 10 B]`,
+			Expect(logs.AllUntimed()).To(ContainElement(
+				observer.LoggedEntry{
+					Entry: zapcore.Entry{
+						Level:   zapcore.InfoLevel,
+						Message: `call "<method>"`,
+					},
+					Context: []zapcore.Field{
+						zap.Int("param_size", 2),
+						zap.Int("result_size", 10),
+					},
 				},
 			))
 		})
@@ -135,9 +145,15 @@ var _ = Describe("func Exchange() (single request)", func() {
 			)
 
 			Expect(err).To(MatchError("<write error>"))
-			Expect(buffer.Messages()).To(ContainElement(
-				logging.BufferedLogMessage{
-					Message: `unable to write JSON-RPC response: <write error>`,
+			Expect(logs.AllUntimed()).To(ContainElement(
+				observer.LoggedEntry{
+					Entry: zapcore.Entry{
+						Level:   zapcore.ErrorLevel,
+						Message: `unable to write JSON-RPC response`,
+					},
+					Context: []zapcore.Field{
+						zap.String("error", "<write error>"),
+					},
 				},
 			))
 		})
@@ -167,9 +183,15 @@ var _ = Describe("func Exchange() (single request)", func() {
 
 			Expect(err).ShouldNot(HaveOccurred())
 			Expect(called).To(BeTrue())
-			Expect(buffer.Messages()).To(ContainElement(
-				logging.BufferedLogMessage{
-					Message: `notify "<method>" [params: 2 B]`,
+			Expect(logs.AllUntimed()).To(ContainElement(
+				observer.LoggedEntry{
+					Entry: zapcore.Entry{
+						Level:   zapcore.InfoLevel,
+						Message: `notify "<method>"`,
+					},
+					Context: []zapcore.Field{
+						zap.Int("param_size", 2),
+					},
 				},
 			))
 		})